+++ pkg/v1/task/task.go
@@ -189,7 +189,7 @@ func makeContractTaskOpts(t *TaskShared, taskDefaults *create.TaskDefaults) *con
 			concurrencyOpts.Name = &name
 		}
 
-		if concurrency.TenantScoped {
+		if concurrency.IsTenantScoped {
 			tenantScoped := true
 			concurrencyOpts.IsTenantScoped = &tenantScoped
 		}